requirements.txt
bertopic==0.16.0
gensim==4.3.2
hdbscan==0.8.33
huggingface-hub==0.20.3
ipykernel==6.29.2
ipython==8.18.1
matplotlib==3.8.3
nltk==3.8.1
numpy==1.23.1
pandas==2.2.0
scikit-learn==1.4.1.post1
sentence-transformers==2.3.1
transformers==4.37.2
umap-learn==0.5.5
urllib3==2.2.1